pyproject.toml
[project]
name = "ruler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "nltk>=3.9.1",
    "pandas>=2.3.2",
    "tqdm>=4.67.1",
    "wikiextractor>=3.0.6",
    "tomli>=2.0.1",
    "PyYAML>=6.0.2",
    "tiktoken>=0.7.0",
    "tenacity>=8.2.3",
    "numpy>=1.26",
    "scipy>=1.11.0",
]
